requirements.txt
Flask==3.0.0
python-dotenv
gunicorn
flask-talisman
whitenoise
Pillow
